==> Notebooks/nb2_ufo_weekly_datecleaning.ipynb <==
%matplotlib inline
import pandas as pd
import numpy as np
from numpy import arange
import math

import seaborn as sns
sns.set_style("white")

import scipy as sp
from scipy import linalg, optimize

import statsmodels.api as sm
from statsmodels.sandbox.regression.predstd import wls_prediction_std

import collections
from collections import Counter

import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.path as mpath
from matplotlib.dates import DateFormatter

import calendar
import datetime as dt
fromtimestamp = dt.datetime.fromtimestamp


# import raw data and confirm
df = pd.read_csv('../Data/Raw/ufo_raw.csv')
df.head(1)

# copy intow(orking)df
wdf = df.copy()
# review columns
wdf.columns

# add col datetype
wdf['datetype'] = pd.to_datetime(wdf['datetime']).dt.date

# add col dtdtype
wdf['dttyp'] = pd.to_datetime(wdf['datetype'])

# add col weekofyear
wdf['weekofyear'] = wdf['dttyp'].dt.isocalendar().week

# add col year
wdf['year'] = wdf['dttyp'].dt.year

# add col year_week
wdf['year_week'] = wdf['year'].astype(str) + "-" + wdf['weekofyear'].astype(str)

# drop useless cols and update wdf
wdf = wdf.drop(columns=['Unnamed: 7', 'Unnamed: 8', 'Unnamed: 9', 'Unnamed: 10', 'datetype','datetime'])

# new df "ufodata"
ufodata = pd.DataFrame(wdf['year_week'])

# add col date_of_sight
ufodata['date_of_sight'] = wdf['dttyp']

# add cols for year, month, week_of_year,state,city,sight_summary, shape
ufodata['year'] = wdf['year']
ufodata['month'] = wdf['dttyp'].dt.month
ufodata['week_of_year'] = wdf['weekofyear']
ufodata['state'] = wdf['State']
ufodata['city'] = wdf['City']
ufodata['sight_summary'] = wdf['Summary']
ufodata['ufo_shape'] = wdf['Shape']

# confirm
ufodata.head()

# make new df "datedf" with daterange

dt = pd.date_range("20060102","20200906", freq='W')
datedf = pd.DataFrame(dt).rename(columns={0:'startdayweek'})

# add cols weeknum,month,year,week_id
datedf['weeknum'] = datedf['startdayweek'].dt.isocalendar().week
datedf['month'] = datedf['startdayweek'].dt.month
datedf['year'] = datedf['startdayweek'].dt.year
datedf['week_id'] = datedf['year'].astype(str)+"-"+datedf['weeknum'].astype(str)

# make new df ufocount from ufodata.copy, drop cols we don't need
ufocount = ufodata.copy().drop(columns=['index','state','city','sight_summary','ufo_shape']).rename(columns={'year_week':'week_id'} ).sort_values('week_id').reset_index()

# merge date df and count df to ufo_cleaned_fin
ufo_cleaned_fin = pd.merge(left=datedf, right=ufocount, left_on='week_id', right_on='week_id')

# review
ufo_cleaned_fin.head()

# copy to wdfweek df and drop cols we do not need
wdfweek = ufo_cleaned_fin.copy().drop(columns=['weeknum','index','year_x','month_x'])

# store counts per week_id in "ufosperweek" df
ufosperweek = pd.DataFrame(wdfweek['startdayweek'].value_counts()).reset_index().rename(columns={'index':'startdateweek','startdayweek': 'weekly_ufocount'})
# review
ufosperweek.head(1)

# set max count to 100 percent
maxalltime = ufosperweek.weekly_ufocount.max() 

# store val for one percent
onepct = maxalltime/100

# new col with relative percentage vals
ufosperweek['rel_val_alltime'] = ufosperweek['weekly_ufocount']/onepct

# add all cols we want
ufosperweek['year'] = ufosperweek['startdateweek'].dt.year
ufosperweek['month'] = ufosperweek['startdateweek'].dt.month
ufosperweek['weeknum'] = ufosperweek['startdateweek'].dt.isocalendar().week 
ufosperweek['week_id'] = ufosperweek['year'].astype(str) + "-" + ufosperweek['weeknum'].astype(str)

# review
ufosperweek.head()

# export csv and set to comment after
#ufosperweek.to_csv('../Data/Datasets/cleandate_weekly_reval.csv')
