==> Notebooks/nb4_google_data_concatenate.ipynb <==
%matplotlib inline
import pandas as pd
import numpy as np
from numpy import arange
import math

import seaborn as sns
sns.set_style("white")

import scipy as sp
from scipy import linalg, optimize

import statsmodels.api as sm
from statsmodels.sandbox.regression.predstd import wls_prediction_std

import collections
from collections import Counter

import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.path as mpath
from matplotlib.dates import DateFormatter

import calendar
import datetime as dt
fromtimestamp = dt.datetime.fromtimestamp


# import and review ufo df
ufo = pd.read_csv('../data/ufo_clean.csv', index_col = 0)
ufo = ufo.rename(columns={'index':'ufo_id_num'})
# start ufo_id index 1
ufo['ufo_id_num'] += 1
# review
ufo.head(1)

# empty storages for google df data
googledfs = []
googledfnames = []
googledfdict = dict(zip(googledfnames,googledfs))

# trump import, rename, sorting, assign topic
trump = pd.read_csv('../Data/Raw/Trump.csv',header=1).sort_values('Week').rename(columns={'Donald Trump: (United States)':'rel_val'})
trump['topic']="trump"
googledfs.append(trump)
googledfnames.append('trump')
trump.head(1)

# obama import, rename, sorting, assign topic
obama = pd.read_csv('../Data/Raw/Obama.csv',header=1).sort_values('Week').rename(columns={'Barack Obama: (United States)':'rel_val'})
obama['topic']="obama"
googledfs.append(obama)
googledfnames.append('obama')
obama.head(1)

# kremlin  import, rename, sorting, assign topic
kremlin =  pd.read_csv('../Data/Raw/Kremlin.csv',header=1).sort_values('Week').rename(columns={'Kremlin: (United States)':'rel_val'})
kremlin['topic']="kremlin"
googledfs.append(kremlin)
googledfnames.append('kremlin')
kremlin.head(1)

# ebola import, rename, sorting, assign topic
ebola =  pd.read_csv('../Data/Raw/Ebola.csv',header=1).sort_values('Week').rename(columns={'Ebola: (United States)':'rel_val'})
ebola['topic']="ebola"
googledfs.append(ebola)
googledfnames.append('ebola')
ebola.head(1)

# corona import, rename, sorting, assign topic
corona =  pd.read_csv('../Data/Raw/Corona.csv',header=1).sort_values('Week').rename(columns={'Coronavirus: (United States)':'rel_val'})
corona['topic']="corona"
googledfnames.append('corona')
googledfs.append(corona)
corona.head(1)

# import arab spring, rename, sort, assign topic
arab = (pd.read_csv('../Data/Raw/arabs_2010_2011_googlesearch.csv')
        .rename(columns={'arab spring':'rel_val' ,'date' : 'Week'})
        .sort_values('Week'))
arab = arab.drop(['isPartial'], axis = 1)
arab['topic']="arab_spring"
googledfs.append(arab)
googledfnames.append('arab')
arab.head(1)

# import boston marathon, rename, sort, assign topic
boston = (pd.read_csv('../Data/Raw/boston_marathon_weeks2013_googlesearch.csv')
            .rename(columns={'boston marathon bombings':'rel_val' ,'date' : 'Week'})
             .sort_values('Week'))
boston = boston.drop(['isPartial'], axis = 1)
boston['topic']="boston"
googledfs.append(boston)
googledfnames.append('boston')
boston.head(1)

# import financial crisis, rename, sort, assign topic
fcrisis = (pd.read_csv('../Data/Raw/fcrisis_weeks2008_googlesearch.csv')
           .rename(columns={'financial crisis':'rel_val' ,'date' : 'Week'})
           .sort_values('Week'))
fcrisis = fcrisis.drop(['isPartial'], axis = 1)
fcrisis['topic']="fcrisis"
googledfs.append(fcrisis)
googledfnames.append('fcrisis')
fcrisis.head(1)

googledfs

# assign datetime col to all dfs
for df in googledfs:
    df['week_dt'] = pd.to_datetime(df['Week'])

# confirm
corona.head(1)

# concat all dfs to "alldfs"
alldfs = pd.concat(googledfs).drop(columns='Week')

# assign cols "year","month","week_of_year"
alldfs['year'] = alldfs['week_dt'].dt.year.astype(str)
alldfs['month'] = alldfs['week_dt'].dt.month.astype(str)
alldfs['week_of_year'] = alldfs['week_dt'].dt.isocalendar().week

#review
alldfs.head()

# assign "week_id" col
alldfs['week_id'] = alldfs['year'] + "-" + alldfs['week_of_year'].astype(str)

# confirm
alldfs.head()

# changing <1 values in column rel val to 0.5 , changing to numeric dtype
alldfs['rel_val'] = np.where(alldfs['rel_val'] == '<1', 0.5 , alldfs['rel_val'])
alldfs['rel_val'] = pd.to_numeric(alldfs['rel_val'])
alldfs.dtypes

# export to csv, set to comment after export:
#alldfs.to_csv("../Data/Datasets/googletopics_weekid_v3.csv")
