==> Notebooks/nb8_OLS_ufos.ipynb <==
# import usual suspects

%matplotlib inline
import pandas as pd
import numpy as np
from numpy import arange
import math

import seaborn as sns
sns.set_style("white")

import scipy as sp
from scipy import linalg, optimize

import statsmodels.api as sm
from statsmodels.sandbox.regression.predstd import wls_prediction_std
from statsmodels.tsa.ar_model import AR
from statsmodels.tsa.stattools import adfuller
from statsmodels.tsa.arima_model import ARMA

import collections
from collections import Counter

import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.path as mpath
from matplotlib.dates import DateFormatter

import calendar
import datetime as dt
fromtimestamp = dt.datetime.fromtimestamp
from datetime import datetime

from pandas.plotting import autocorrelation_plot
from pandas.plotting import lag_plot

# import cleaned weekly ufo data

ufo = pd.read_csv('../Data/Datasets/cleandate_weekly_reval.csv',index_col=0)
# copy into w(orking)ufo df
wufo = ufo.copy().rename(columns={'rel_val_alltime':'weekly_ufopct'})
# confirm
wufo.head(1)

# add date index

# add date col and sort values
wufo['date'] = ufo['startdateweek']
wufo = wufo.sort_values('date')
# add constant
wufo['const'] = 1
# make date col datetime and set index to date
wufo.date = pd.to_datetime(wufo.date)
# dateindex
wufo.index = pd.DatetimeIndex(wufo.date, freq='W')
# confirm
wufo.columns

# add numeric bool col for corona timeframe

row_indexes = [x for x in wufo[wufo['startdateweek']>'2019-09-01'].index]
wufo['corona'] =1
wufo.loc[row_indexes,'corona']=0
wufo.head()

# review general correlations

wufo.corr()

######################## ######################## ######################## ######################## ########################
######################### ######################## ######################## ######################## ########################
#
# review standard deviations
#
print('\n\nstandard deviations UFO Count per Week:',(" "*20), "Total Count:",(" "*20),"Std in % of Total: \n", (" "*30),wufo.weekly_ufocount.std()," "*20,wufo.weekly_ufocount.sum()," "*20, wufo.weekly_ufocount.std()/(wufo.weekly_ufocount.sum()/100))
print('\n\nstandard deviations UFO Percentage per Week:',"\n"," "*30,wufo.weekly_ufopct.std())

######################## ######################## ######################## ######################## ########################
######################### ######################## ######################## ######################## ########################

# adding daynum col

wufo['daynum'] = wufo['date'].dt.dayofyear

# OLS: all time date related vals ⇔ ufos

results = sm.OLS(wufo["weekly_ufocount"], 
                  wufo[['const','daynum','month', 'year','corona']]).fit()
    # summary
results.summary()

##################################################################################
##################################################################################
### NOTES:
#
#       highest coef: month (11.5737)
#                 
##################################################################################
##################################################################################


# OLS: only month vals ⇔ ufos

results = sm.OLS(wufo["weekly_ufocount"], 
                  wufo[['const','month']]).fit()
    # summary
results.summary()

##################################################################################
##################################################################################
### NOTES:
#
#              Adjusted R-squared month: 0.043
#
##################################################################################
##################################################################################

# OLS: only year vals ⇔ ufos

results = sm.OLS(wufo["weekly_ufocount"], 
                  wufo[['const','year']]).fit()
    # summary
results.summary()

##################################################################################
##################################################################################
### NOTES:
#
#       Adjusted R-squared year: 0.031
#             
##################################################################################
##################################################################################


##################################################################################
#
# OLS: monthly mean corona/before ⇔ ufos

# import dataset monthly means
tsdf = pd.read_csv('../Data/Datasets/monmeans_corona.csv',index_col=0)

# add constant and numeric bool corona
tsdf['const'] = 1
tsdf['corona'] = 1
tsdf = tsdf.loc[tsdf['Timeframe:'] =='UFOs before corona (monthly mean)'].assign(corona=0)

# OLS: month, corona, before ⇔ ufos

results = sm.OLS(tsdf["UFOs"], 
                  tsdf[['const','Month','corona']]).fit()
    # summary
results.summary()

##################################################################################
##################################################################################
