==> Notebooks/nb11_Ttests.ipynb <==
# import usual suspects

%matplotlib inline
import pandas as pd
import numpy as np
from numpy import arange
import math

import seaborn as sns
sns.set_style("white")

import scipy as sp
from scipy import linalg, optimize

import statsmodels.api as sm
from statsmodels.sandbox.regression.predstd import wls_prediction_std
from statsmodels.tsa.ar_model import AR
from statsmodels.tsa.stattools import adfuller
from statsmodels.tsa.arima_model import ARMA

import collections
from collections import Counter

import matplotlib as mpl
import matplotlib.pyplot as plt
import matplotlib.path as mpath
from matplotlib.dates import DateFormatter

import calendar
import datetime as dt
fromtimestamp = dt.datetime.fromtimestamp
from datetime import datetime

from pandas.plotting import autocorrelation_plot
from pandas.plotting import lag_plot

######################## ######################## ######################## ######################## ########################
######################## ######################## ######################## ######################## ########################
#
# import cleaned weekly ufo data

ufo = pd.read_csv('../Data/Datasets/weeklufo_corona.csv',index_col=0)

# copy into w(orking)ufo df
wufo = ufo.copy().rename(columns={'rel_val_alltime':'weekly_ufopct'})

# confirm
wufo.tail()

# import dataset ufo arma + copy to ufa df

ufoarma = pd.read_csv('../Data/Datasets/nb6_ufoarma.csv')
ufa = ufoarma.copy()

# storing in lists
obs1 = ufa['observed'].to_list()
pred1 = ufa['predicted'].to_list()
dat = ufa['date'].to_list()

# datetime col + index for ufa df
ufa.date = pd.to_datetime(ufa.date)
ufa.index = pd.DatetimeIndex(ufa.date,freq='W')

# t(emporary) ufa copy without date col
tufa = ufa.copy().drop(columns={'date'})
tufa.head()

# storing in lists
obs2 = tufa['observed'].to_list()
pred2 = tufa['predicted'].to_list()

# T-Tests

# T-test 1 sample: weekly ufo counts

        # storing ufo counts to list
ufocounts = wufo['weekly_ufocount'].to_list()

        # perform ttest
sp.stats.ttest_1samp(ufocounts, popmean=0)
             # scipy.stats.ttest_1samp accepts the population mean as the second "group"

    
                # OUT: Ttest_1sampResult(statistic=52.504527546746914, pvalue=7.577308488245331e-256)   
        
                                                                        

 # T-test 2 sample: 
    # T-tests for predicted and observerd val during corona 


    # 1. t-test two samples, df: ufa

sp.stats.ttest_ind(obs1,pred1)

    # OUT: Ttest_indResult(statistic=0.36349723616468227, pvalue=0.7172168241649526)
    
                                                                         

# Summary T-Tests

T-Test can be used to **compare the means of two samples**. 

If you have more than 2 samples, you will have to run a pairwise T-Test on all samples - or use another and ***muuuuuch*** better statistical hypothesis method called **Anova**. 

Ttest is of more use if you don’t know the population’s mean and standard deviation. But we do.

Plus: In the T-Test, you are comparing **2 samples of an unknown population**. A sample is a randomly chosen set of data points from a population. If you do know the population’s mean and standard deviation, you would actually run a **Z-Test** instead.

**So. Ttest is actually only for small number of samples. Not huge ones like our impressive UFO dataset.** 

The T-Test is commonly used when you have less than 30 samples in each of the groups. We can use it for the monthly means before and after corona. 

**This⬈ is DONE in Notebook7 of this project**

T-Test is only used if the samples are SMALL or follow a NORMAL distribution. Why? Ask the **Central Limit Theorem**.
